refactor: Vote Progress, Result 파일 이동 및 props 변경
