+++ src/app/result/[postId]/_components/VoteBlock/VoteResult.tsx
